refactor(agent-runtime): replace dead litter-bridge gateway with metadata publisher
The litter-bridge HTTP gateway (/api/litter-bridge/v1) was Design A: a
signed, versioned bridge that forwarded controller-status, session-read,
and agent-turn requests from the KittyLitter daemon. Design B replaced
it with a scoped PiBridge on the daemon side (alleycat f443a83), and the
alleycat HTTP client that called this endpoint was deleted in alleycat
PR #34. The gateway was left running on 127.0.0.1:8081, used by nobody.

The gateway did two things; only one is still live:

1. Served /api/litter-bridge/v1 (dead — no caller remains).
2. Wrote litter-bridge.json publishing piAgentDir and piRuntime so the
   KittyLitter daemon can launch the same Pi version and data dir this
   instance owns (live — alleycat reads it via load_gateway_metadata).

Extract the live metadata publishing into a focused 150-line module
(litter-bridge-metadata.ts) and delete the rest:

- services/agent-runtime/src/litter-bridge-gateway.ts (3,346 lines)
- services/agent-runtime/src/litter-bridge-mutation-ledger.ts (571 lines)
- services/agent-runtime/test/litter-bridge-gateway.test.ts (1,596 lines)
- services/agent-runtime/test/litter-bridge-mutation-ledger.test.ts (229 lines)
- shared/agent/litter-bridge.ts (700 lines) — only imported by the gateway
- shared/agent/litter-bridge.test.ts (526 lines)

server.ts drops the HTTP route and calls the new publisher's
publishMetadata/dispose, which write and clean up the same
litter-bridge.json with the same schema.

Verified:
- services/agent-runtime: bun test 25 pass, 0 fail.
- services/agent-runtime: tsc --noEmit clean.
- npm run check:contracts: passed.
- npm run check:structure: passed.
