+++ services/agent-runtime/src/server.ts
@@ -48,7 +48,7 @@ import {
   handlePtyResize,
   handlePtyStream,
 } from "./http/pty-handlers";
-import { createLitterBridgeGateway } from "./litter-bridge-gateway";
+import { createLitterBridgeMetadataPublisher } from "./litter-bridge-metadata";
 import { handleAgentModels } from "./http/model-handlers";
 import {
   handleAllSessions,
@@ -62,13 +62,11 @@ markAgentRuntimeProcess();
 startAutomationScheduler();
 
 const app = new Hono();
-const litterBridgeGateway = createLitterBridgeGateway();
+const litterBridgeMetadata = createLitterBridgeMetadataPublisher();
 
 app.get("/health", (c) =>
   c.json({ ok: true, service: "local-studio-agent-runtime", pid: process.pid }),
 );
-app.post("/api/litter-bridge/v1", (c) => litterBridgeGateway.handle(c.req.raw));
-
 app.post("/api/agent/turn", (c) => handleAgentTurn(c.req.raw));
 app.post("/api/agent/abort", (c) => handleAgentAbort(c.req.raw));
 app.post("/api/agent/compact", (c) => handleAgentCompact(c.req.raw));
@@ -131,12 +129,12 @@ app.post("/api/agent/browser/:verb", (c) => handleBrowserVerb(c.req.raw, c.req.p
 const port = Number(process.env.PORT) > 0 ? Number(process.env.PORT) : 8081;
 
 serve({ fetch: app.fetch, port, hostname: "127.0.0.1" }, (info) => {
-  litterBridgeGateway.publishMetadata(info.port);
+  litterBridgeMetadata.publishMetadata(info.port);
   console.log(
     `[agent-runtime] listening on http://127.0.0.1:${info.port} (pid ${process.pid}, node ${process.version})`,
   );
 });
 
-process.once("exit", () => litterBridgeGateway.dispose());
+process.once("exit", () => litterBridgeMetadata.dispose());
 process.once("SIGINT", () => process.exit(0));
 process.once("SIGTERM", () => process.exit(0));
